각 context 및 컴포넌트 분리 완료
