+++ src/contexts/notificationContext.ts
@@ -1,0 +1,22 @@
+import { Notification } from "../type";
+import { createContext, useContext } from "react";
+
+export interface NotificationContextType {
+  notifications: Notification[];
+  addNotification: (message: string, type: Notification["type"]) => void;
+  removeNotification: (id: number) => void;
+}
+
+export const NotificationContext = createContext<
+  NotificationContextType | undefined
+>(undefined);
+
+export const useNotificationContext = () => {
+  const context = useContext(NotificationContext);
+  if (context === undefined) {
+    throw new Error(
+      "useNotificationContext must be used within an NotificationProvider",
+    );
+  }
+  return context;
+};
