+++ src/contexts/userContext.ts
@@ -1,0 +1,20 @@
+import { User } from "../type";
+import { createContext, useContext } from "react";
+
+export interface UserContextType {
+  user: User | null;
+  login: (email: string, password: string) => void;
+  logout: () => void;
+}
+
+export const UserContext = createContext<UserContextType | undefined>(
+  undefined,
+);
+
+export const useUserContext = () => {
+  const context = useContext(UserContext);
+  if (context === undefined) {
+    throw new Error("useUserContext must be used within an UserProvider");
+  }
+  return context;
+};
